Added ability to look at and create compounds
